requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.6
flask-jwt-extended==4.4.4
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.28.1
flask-bcrypt==1.0.1
flask-cors==3.0.10
python-dotenv==1.0.0
flasgger==0.9.5
email-validator==2.0.0
flask-migrate==3.1.0
gunicorn==20.1.0
pytest==7.3.1
sqlalchemy==1.4.41
werkzeug==2.3.4
flask-talisman==1.1.0  # For additional security headers
blinker==1.6.2  # Dependency for signals
